MOre refactoring and modulation'
