+++ src/components/GroupList/index.tsx
@@ -1,6 +1,6 @@
 import './GroupList.css'
 import { useState } from 'react'
-import { getChartData } from '../../sagas/actions'
+import { downloadChartData } from '../../store/OpenedInstruments/actions'
 import { useAppDispatch } from '../../app/hooks'
 import { instrumentInfo } from '../../store/OpenedInstruments/types'
 
@@ -11,7 +11,7 @@ const GroupList: React.FC<{
   const dispatch = useAppDispatch()
   const [isOpen, setIsOpen] = useState(false)
 
-  const switchInsturment = (symbol: string) => dispatch(getChartData(symbol))
+  const switchInsturment = (symbol: string) => dispatch(downloadChartData(symbol))
 
   return (
     <div className="group-list">
