+++ src/components/Statusbar/index.tsx
@@ -1,11 +1,11 @@
 import { useAppDispatch, useAppSelector } from '../../app/hooks'
-import { WebSocketDisconnect } from '../../sagas/actions'
+import { LogoutUser } from '../../store/LoginData/actions'
 import { getUserId } from '../../store/LoginData/selectors'
 
 const Statusbar = () => {
   const dispatch = useAppDispatch()
   const UserID = useAppSelector(getUserId)
-  const logout = () => dispatch(WebSocketDisconnect())
+  const logout = () => dispatch(LogoutUser())
   return (
     <>
       <div>UserID: {UserID}</div>
