+++ src/sagas/MainSaga.ts
@@ -1,10 +1,7 @@
-import {fork, all} from "redux-saga/effects";
-import WebsocketSaga from './WebsocketSaga'
-
+import { fork, all } from 'redux-saga/effects'
+import ConnectUserSaga from '../store/LoginData/saga'
+import AllSagaWatcher from '../store/MainConnection/saga'
 
 export default function* MainSaga() {
-	yield all([
-		fork(WebsocketSaga),
-	])
-
+  yield all([fork(ConnectUserSaga), fork(AllSagaWatcher)])
 }
