--- src/sagas/types/index.ts
@@ -1,11 +1,0 @@
-
-export const enum WSACTIONS {
-	connect = "WebSocket Connect",
-	connectStream = "Websocket Stream Connect",
-	getChartData = "get Main chart data",
-	disconnect = "WebSocket Disconnect",
-	saveChartData = "SaveChartData",
-	getAccountData = "get account Data",
-	passAccountData = "pass Account Data"
-
-}
