+++ src/store/MainConnection/reducers.ts
@@ -1,0 +1,2 @@
+const reducers = {}
+export default reducers
