+++ src/store/OpenedInstruments/actions.ts
@@ -1,0 +1,11 @@
+import { WSACTIONS } from '../LoginData/types'
+
+export const downloadChartData = (
+  symbol: string,
+): {
+  type: WSACTIONS
+  payload: string
+} => ({
+  type: WSACTIONS.downloadChartData,
+  payload: symbol,
+})
