+++ src/utils/websocket/index.ts
@@ -1,16 +1,9 @@
-import {wsRequest} from "../../types/RequestResponseTypes"
 import handleResponse from './mainSessionHandler'
-import {balanceStreamHandlers} from './balanceHandler'
-
+import { balanceStreamHandlers } from './balanceHandler'
+import { wsRequest } from '../../store/LoginData/types'
 
 export const send = (ws: WebSocket, msg: wsRequest): void => {
-	ws.send(JSON.stringify(msg))
+  ws.send(JSON.stringify(msg))
 }
 
-export {
-	handleResponse,
-	balanceStreamHandlers
-}
-
-
-
+export { handleResponse, balanceStreamHandlers }
